+++ picopt/handlers/gif.py
@@ -6,15 +6,15 @@ from typing import Optional
 from PIL import Image
 from PIL.GifImagePlugin import GifImageFile
 
-from picopt.handlers.handler import Format
+from picopt.handlers.handler import FileFormat
 from picopt.handlers.image import ImageHandler
 
 
 class Gif(ImageHandler):
     """GIF handler."""
 
-    OUTPUT_FORMAT = GifImageFile.format
-    OUTPUT_FORMAT_OBJ = Format(OUTPUT_FORMAT, True, False)
+    OUTPUT_FORMAT_STR = GifImageFile.format
+    OUTPUT_FILE_FORMAT = FileFormat(OUTPUT_FORMAT_STR, True, False)
     PROGRAMS: dict[str, Optional[str]] = ImageHandler.init_programs(
         ("gifsicle", "pil2gif")
     )
@@ -30,14 +30,14 @@ class Gif(ImageHandler):
             return old_path
 
         shutil.copy2(old_path, new_path)
-        args = tuple(self._ARGS_PREFIX + [str(new_path)])
+        args = (*self._ARGS_PREFIX, str(new_path))
         self.run_ext(args)
         return new_path
 
     def pil2gif(self, old_path: Path, new_path: Path) -> Path:
         """Pillow gif optimization."""
         with Image.open(old_path) as image:
-            image.save(new_path, self.OUTPUT_FORMAT, optimize=True, save_all=True)
+            image.save(new_path, self.OUTPUT_FORMAT_STR, optimize=True, save_all=True)
         image.close()  # for animated images
         return new_path
 
@@ -45,4 +45,4 @@ class Gif(ImageHandler):
 class AnimatedGif(Gif):
     """Animated GIF handler."""
 
-    OUTPUT_FORMAT_OBJ = Format(Gif.OUTPUT_FORMAT, True, True)
+    OUTPUT_FILE_FORMAT = FileFormat(Gif.OUTPUT_FORMAT_STR, True, True)