+++ tests/integration/test_old_timestamps.py
@@ -2,7 +2,7 @@
 import shutil
 
 from picopt import PROGRAM_NAME, cli
-from picopt.old_timestamps import _OLD_TIMESTAMPS_NAME
+from picopt.old_timestamps import OLD_TIMESTAMPS_NAME
 from tests import IMAGES_DIR, get_test_dir
 
 __all__ = ()
@@ -44,12 +44,11 @@ class TestContainersDir:
 
     def teardown_method(self) -> None:
         """Tear down method."""
-        if TMP_ROOT.exists():
-            shutil.rmtree(TMP_ROOT)
+        shutil.rmtree(TMP_ROOT, ignore_errors=True)
 
     def test_old_timestamp_same_dir(self) -> None:
         """Test old timestamp in same dir."""
-        old_ts_path = TMP_ROOT / _OLD_TIMESTAMPS_NAME
+        old_ts_path = TMP_ROOT / OLD_TIMESTAMPS_NAME
         old_ts_path.touch()
         args = (PROGRAM_NAME, "-rtvv", str(TMP_ROOT))
         res = cli.main(args)
@@ -61,24 +60,24 @@ class TestContainersDir:
         """Test old timestamp child."""
         child_root = TMP_ROOT / "child"
         child_root.mkdir(exist_ok=True)
-        shutil.copy(SRC_JPG, child_root)
-        old_ts_path = child_root / _OLD_TIMESTAMPS_NAME
+        shutil.move(TMP_ROOT / FN, child_root)
+        old_ts_path = child_root / OLD_TIMESTAMPS_NAME
         old_ts_path.touch()
-        args = (PROGRAM_NAME, "-rtvv", str(child_root))
+        args = (PROGRAM_NAME, "-rtvv", str(TMP_ROOT))
         res = cli.main(args)
         assert res
         assert not old_ts_path.exists()
-        self._assert_sizes(0)
+        self._assert_sizes(0, root=child_root)
 
     def test_old_timestamp_parent(self) -> None:
         """Test old timestamp in parent."""
         child_root = TMP_ROOT / "child"
         child_root.mkdir(exist_ok=True)
-        shutil.copy(SRC_JPG, child_root)
-        old_ts_path = TMP_ROOT / _OLD_TIMESTAMPS_NAME
+        shutil.move(TMP_ROOT / FN, child_root)
+        old_ts_path = TMP_ROOT / OLD_TIMESTAMPS_NAME
         old_ts_path.touch()
         args = (PROGRAM_NAME, "-rtvv", str(child_root))
         res = cli.main(args)
         assert res
         assert old_ts_path.exists()
-        self._assert_sizes(0)
+        self._assert_sizes(0, root=child_root)