+++ tests/unit/test_png_bit_depth.py
@@ -13,13 +13,13 @@ TEST_SRC_PATH_JPG = IMAGES_DIR / "test_jpg.jpg"
 def test_png_bit_depth() -> None:
     """Test PNG bit depth."""
     res = png_bit_depth(TEST_SRC_PATH)
-    assert res == 8
+    assert res == 8  # noqa PLR2004
 
 
 def test_png_bit_depth_16() -> None:
     """Test PNG bit depth 16."""
     res = png_bit_depth(TEST_SRC_PATH_16)
-    assert res == 16
+    assert res == 16  # noqa PLR2004
 
 
 def test_png_bit_depth_invalid() -> None: